% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_maps.R
\name{add_state_clusters}
\alias{add_state_clusters}
\title{Create case clusters at state level.}
\usage{
add_state_clusters(leaflet_map, data, states)
}
\description{
\code{add_state_clusters()} creates clusters for each state level
}
