% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data-triangulation.R
\name{chart_label}
\alias{chart_label}
\title{Create the chart label}
\usage{
chart_label(picker_state_var, picker_lga_var)
}
\arguments{
\item{picker_state_var}{reactive value of the selected State}

\item{picker_lga_var}{reactive value of the selected LGA}
}
\value{
a character string
}
\description{
Create the chart label which shows the State:LGA text in the
place of the chart title.
}
