% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data-triangulation.R
\name{disease_list_util}
\alias{disease_list_util}
\title{Create a list of diseases}
\usage{
disease_list_util()
}
\value{
A list
}
\description{
Create a list of diseases in the Dashboard in one place. So that you keep track of the spellings and
easy additional extra diseases in the future
}
