% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data-triangulation.R
\name{font_axis_title}
\alias{font_axis_title}
\title{Style font  of the plot text used in the axises}
\usage{
font_axis_title()
}
\description{
Style font  of the plot text used in the axises
}
