% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data-triangulation.R
\name{font_hoverlabel}
\alias{font_hoverlabel}
\title{Style font  of the text used when a user hovers over a plot}
\usage{
font_hoverlabel()
}
\description{
Style font  of the text used when a user hovers over a plot
}
