% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_measles_page.R, R/mod_meningitis_page.R,
%   R/mod_yellow_fever_page.R
\name{measles_page}
\alias{measles_page}
\alias{meningitis_page}
\alias{yellow_fever_page}
\title{Create a Measles/ Yellow Fever/ Meningitis page}
\usage{
measles_page()

meningitis_page()

yellow_fever_page()
}
\value{
A Brochure Page
}
\description{
\code{measles_page()} creates the measles page

\code{yellow_fever_page()} creates the yellow fever page

\code{meningitis_page()} creates the meningitis page
}
\details{
Since the output is a Brochure page that means the ui and server function are defined in the page
and hence contains code for the UI and server of the page
}
