% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_age_group_of_confirmed_measles_cases_by_vaccination_status.R,
%   R/mod_age_group_of_confirmed_meningitis_cases_by_vaccination_status.R,
%   R/mod_age_group_of_confirmed_yellow_fever_cases_by_vaccination_status.R
\name{mod_age_group_of_confirmed_measles_cases_by_vaccination_status_server}
\alias{mod_age_group_of_confirmed_measles_cases_by_vaccination_status_server}
\alias{mod_age_group_of_confirmed_meningitis_cases_by_vaccination_status_server}
\alias{mod_age_group_of_confirmed_yellow_fever_cases_by_vaccination_status_server}
\title{Create shiny module server side for chart 2}
\usage{
mod_age_group_of_confirmed_measles_cases_by_vaccination_status_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)

mod_age_group_of_confirmed_meningitis_cases_by_vaccination_status_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)

mod_age_group_of_confirmed_yellow_fever_cases_by_vaccination_status_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)
}
\arguments{
\item{id}{The id (string) to be used  for name-spacing in the module.}

\item{picker_year_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_month_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_state_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_lga_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}
}
\value{
A shiny server Module
}
\description{
\code{mod_age_group_of_confirmed_measles_cases_by_vaccination_status_server()}
Creates a shiny Module for the server of confirmed age group of disease cases vaccination status
currently chart 2
}
