% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_age_group_of_confirmed_measles_cases_by_vaccination_status.R,
%   R/mod_age_group_of_confirmed_meningitis_cases_by_vaccination_status.R,
%   R/mod_age_group_of_confirmed_yellow_fever_cases_by_vaccination_status.R
\name{mod_age_group_of_confirmed_measles_cases_by_vaccination_status_ui}
\alias{mod_age_group_of_confirmed_measles_cases_by_vaccination_status_ui}
\alias{mod_age_group_of_confirmed_meningitis_cases_by_vaccination_status_ui}
\alias{mod_age_group_of_confirmed_yellow_fever_cases_by_vaccination_status_ui}
\title{Create shiny module UI side for chart 2}
\usage{
mod_age_group_of_confirmed_measles_cases_by_vaccination_status_ui(id)

mod_age_group_of_confirmed_meningitis_cases_by_vaccination_status_ui(id)

mod_age_group_of_confirmed_yellow_fever_cases_by_vaccination_status_ui(id)
}
\arguments{
\item{id}{The id (string) to be used  for name-spacing in the module.}
}
\value{
A shiny UI Module
}
\description{
\code{mod_age_group_of_confirmed_measles_cases_by_vaccination_status_ui()}
Creates a shiny Module UI of confirmed age group of disease cases vaccination status
currently chart 2
}
