% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_confirmed_measles_cases_MCV1_coverage.R,
%   R/mod_meningitis_coverage_confirmed_cases.R,
%   R/mod_yf_coverage_confirmed_cases.R
\name{mod_confirmed_measles_cases_MCV1_coverage_server}
\alias{mod_confirmed_measles_cases_MCV1_coverage_server}
\alias{mod_meningitis_coverage_confirmed_cases_server}
\alias{mod_yf_coverage_confirmed_cases_server}
\title{Create shiny module server side for chart 1}
\usage{
mod_confirmed_measles_cases_MCV1_coverage_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)

mod_meningitis_coverage_confirmed_cases_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)

mod_yf_coverage_confirmed_cases_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{picker_year_var, picker_month_var, picker_state_var, picker_lga_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}
}
\value{
A shiny server Module
}
\description{
\code{mod_confirmed_measles_cases_MCV1_coverage_server()}
Creates a shiny Module for the server of Confirmed Measles cases,MCV 1 coverage,Alt denominator
currently chart 1
}
