% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_confirmed_measles_cases_MCV1_coverage.R,
%   R/mod_meningitis_coverage_confirmed_cases.R,
%   R/mod_yf_coverage_confirmed_cases.R
\name{mod_confirmed_measles_cases_MCV1_coverage_ui}
\alias{mod_confirmed_measles_cases_MCV1_coverage_ui}
\alias{mod_meningitis_coverage_confirmed_cases_ui}
\alias{mod_yf_coverage_confirmed_cases_ui}
\title{Create shiny module UI side for chart 1}
\usage{
mod_confirmed_measles_cases_MCV1_coverage_ui(id)

mod_meningitis_coverage_confirmed_cases_ui(id)

mod_yf_coverage_confirmed_cases_ui(id)
}
\value{
A shiny UI Module
}
\description{
\code{mod_confirmed_measles_cases_MCV1_coverage_ui()}
Creates a shiny Module UI of Confirmed Measles cases,MCV 1 coverage,Alt denominator
currently chart 1
}
