% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_discrepancy_mcv1_men_A.R,
%   R/mod_discrepancy_mcv1_yellow_fever_given_by_state.R
\name{mod_discrepancy_mcv1_men_A_server}
\alias{mod_discrepancy_mcv1_men_A_server}
\alias{mod_discrepancy_mcv1_yellow_fever_given_by_state_server}
\title{Create shiny module server side for chart 4}
\usage{
mod_discrepancy_mcv1_men_A_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)

mod_discrepancy_mcv1_yellow_fever_given_by_state_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{picker_year_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_month_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_state_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_lga_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}
}
\value{
A shiny server Module
}
\description{
\code{mod_discrepancy_mcv1_yellow_fever_given_by_state_server()}
Creates a shiny Module for the server of Co-administered antigen discrepancy: MCV 1 & Yellow Fever given
currently chart 4
}
