% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_discrepancy_mcv1_men_A.R,
%   R/mod_discrepancy_mcv1_yellow_fever_given_by_state.R
\name{mod_discrepancy_mcv1_men_A_ui}
\alias{mod_discrepancy_mcv1_men_A_ui}
\alias{mod_discrepancy_mcv1_yellow_fever_given_by_state_ui}
\title{Create shiny module UI side for chart 4}
\usage{
mod_discrepancy_mcv1_men_A_ui(id)

mod_discrepancy_mcv1_yellow_fever_given_by_state_ui(id)
}
\arguments{
\item{id}{The id (string) to be used  for name-spacing in the module.}
}
\value{
A shiny UI Module
}
\description{
\code{mod_discrepancy_mcv1_yellow_fever_given_by_state_ui()}
Creates a shiny Module UI of Co-administered antigen discrepancy: MCV 1 & Yellow Fever given
currently chart 4
}
