% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_map_confirmed_measles_cases_mcv1_coverage_annual_data.R,
%   R/mod_map_confirmed_meningitis_cases_coverage_annual_data.R,
%   R/mod_map_confirmed_yellow_fever_cases_yellow_fever_coverage_annual_data.R
\name{mod_map_confirmed_measles_cases_mcv1_coverage_annual_data_ui}
\alias{mod_map_confirmed_measles_cases_mcv1_coverage_annual_data_ui}
\alias{mod_map_confirmed_meningitis_cases_coverage_annual_data_ui}
\alias{mod_map_confirmed_yellow_fever_cases_yellow_fever_coverage_annual_data_ui}
\title{Create shiny module UI side for Chart 7 for Measles and chart 5 (Yellow and Meningitis )}
\usage{
mod_map_confirmed_measles_cases_mcv1_coverage_annual_data_ui(id)

mod_map_confirmed_meningitis_cases_coverage_annual_data_ui(id)

mod_map_confirmed_yellow_fever_cases_yellow_fever_coverage_annual_data_ui(id)
}
\arguments{
\item{id}{The id (string) to be used  for name-spacing in the module.}
}
\value{
A shiny UI Module
}
\description{
\code{mod_map_confirmed_measles_cases_mcv1_coverage_annual_data_ui()}
Creates a shiny Module UI of map of confirmed Measles cases, MCV 1 coverage
currently chart 7
}
