% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mcv1_mcv2_drop_out_rate_nigeria.R
\name{mod_mcv1_mcv2_drop_out_rate_nigeria_server}
\alias{mod_mcv1_mcv2_drop_out_rate_nigeria_server}
\title{Create shiny module server side for chart 5}
\usage{
mod_mcv1_mcv2_drop_out_rate_nigeria_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)
}
\arguments{
\item{id}{Internal parameters for {shiny}.}

\item{picker_year_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_month_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_state_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}

\item{picker_lga_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}
}
\value{
A shiny server Module
}
\description{
\code{mod_mcv1_mcv2_drop_out_rate_nigeria_server()}
Creates a shiny Module for the server of confirmed age group of disease cases vaccination status
currently chart 5
}
