% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mcv1_mcv2_drop_out_rate_nigeria.R
\name{mod_mcv1_mcv2_drop_out_rate_nigeria_ui}
\alias{mod_mcv1_mcv2_drop_out_rate_nigeria_ui}
\title{Create shiny module UI side for chart 5}
\usage{
mod_mcv1_mcv2_drop_out_rate_nigeria_ui(id)
}
\arguments{
\item{id}{The id (string) to be used  for name-spacing in the module.}
}
\value{
A shiny UI Module
}
\description{
\code{mod_mcv1_mcv2_drop_out_rate_nigeria_ui()}
Creates a shiny Module UI of MCV 1, MCV 2 coverage & drop out rate
currently chart 5
}
