% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_measles_vaccine_stock_analysis_measles_coverage.R,
%   R/mod_meningitis_vaccine_stock_analysis.R,
%   R/mod_yellow_fever_vaccine_stock_analysis_yellow_fever_coverage.R
\name{mod_measles_vaccine_stock_analysis_measles_coverage_ui}
\alias{mod_measles_vaccine_stock_analysis_measles_coverage_ui}
\alias{mod_meningitis_vaccine_stock_analysis_ui}
\alias{mod_yellow_fever_vaccine_stock_analysis_yellow_fever_coverage_ui}
\title{Create shiny module UI side for chart 3}
\usage{
mod_measles_vaccine_stock_analysis_measles_coverage_ui(id)

mod_meningitis_vaccine_stock_analysis_ui(id)

mod_yellow_fever_vaccine_stock_analysis_yellow_fever_coverage_ui(id)
}
\arguments{
\item{id}{The id (string) to be used  for name-spacing in the module.}
}
\value{
A shiny UI Module
}
\description{
\code{mod_measles_vaccine_stock_analysis_measles_coverage_ui()}
Creates a shiny Module UI of MCV stock analysis & MCV (1 & 2) given
currently chart 3
}
