% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_national_measles_coverage_different_sources.R
\name{mod_national_measles_coverage_different_sources_server}
\alias{mod_national_measles_coverage_different_sources_server}
\title{Create shiny module server side for chart 6}
\usage{
mod_national_measles_coverage_different_sources_server(id)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\value{
A shiny server Module
}
\description{
\code{mod_national_measles_coverage_different_sources_server()}
Creates a shiny Module for the server of National MCV coverage (\%) by different sources
currently chart 6
}
