% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data-triangulation.R
\name{nav_links}
\alias{nav_links}
\title{Create the disease filter UI}
\usage{
nav_links(disease)
}
\arguments{
\item{disease}{a string of a disease one of "Measles" or "Yellow Fever"or "Meningitis"}
}
\value{
a shiny.tag of the disease filter
}
\description{
Create the disease filter UI
}
