% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data-triangulation.R
\name{plot_number_range}
\alias{plot_number_range}
\title{Determine the scale range of number y-axis}
\usage{
plot_number_range(min_number, max_number)
}
\arguments{
\item{min_number}{The mininum number in the vector}

\item{max_number}{The maximum number in the vector}
}
\value{
A a vector of 2 numbers
}
\description{
Determine the number y-axis where if its beyond 100\% let the upper limit be
the maximum point in the data plus 50\% of that maximum point
}
\details{
formula used

if(max_number > 100){  max_number+ (max_number/ 2) }

Same rule applies to other min_number but towards the negative side
}
