% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states}
\alias{states}
\title{Nigerian States}
\format{
A data frame with 38 rows and 1 column
\describe{
\item{state_name}{States in Nigeria}
}
}
\usage{
states
}
\description{
A dataset containing the Nigeria's States
}
\keyword{datasets}
