% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{states_gadm_sp_data}
\alias{states_gadm_sp_data}
\title{Nigeria's GADM shapefile}
\format{
A gadm_sp of Nigeria's GADM shapefile
}
\usage{
states_gadm_sp_data
}
\description{
A gadm_sp containing the Nigeria's GADM shapefile from GADMTools::gadm_sp_loadCountries
}
\keyword{datasets}
