% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_maps.R
\name{add_state_clusters}
\alias{add_state_clusters}
\title{Create case clusters at state level.}
\usage{
add_state_clusters(leaflet_map, data, states)
}
\arguments{
\item{leaflet_map}{a leaflet object created by leaflet::addPolygons}

\item{data}{reactive value as data frame of SORMAS case based data}

\item{states}{A vector of states in Nigeria returned by states_vector_util()}
}
\description{
\code{add_state_clusters()} creates clusters for each state level
}
