% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_measles_vaccine_stock_analysis_measles_coverage.R,
%   R/mod_meningitis_vaccine_stock_analysis.R,
%   R/mod_yellow_fever_vaccine_stock_analysis_yellow_fever_coverage.R
\name{mod_measles_vaccine_stock_analysis_measles_coverage_server}
\alias{mod_measles_vaccine_stock_analysis_measles_coverage_server}
\alias{mod_meningitis_vaccine_stock_analysis_server}
\alias{mod_yellow_fever_vaccine_stock_analysis_yellow_fever_coverage_server}
\title{Create shiny module server side for chart 3}
\usage{
mod_measles_vaccine_stock_analysis_measles_coverage_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)

mod_meningitis_vaccine_stock_analysis_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)

mod_yellow_fever_vaccine_stock_analysis_yellow_fever_coverage_server(
  id,
  picker_year_var,
  picker_month_var,
  picker_state_var,
  picker_lga_var
)
}
\arguments{
\item{id}{The id (string) to be used  for name-spacing in the module.}

\item{picker_year_var, picker_month_var, picker_state_var, picker_lga_var}{the selected reactive values from \verb{mod_inputs Module}
for picker_year, picker_month, picker_state, picker_lga respectively}
}
\value{
A shiny server Module
}
\description{
\code{mod_measles_vaccine_stock_analysis_measles_coverage_server()}
Creates a shiny Module for the server of MCV stock analysis & MCV (1 & 2) given
currently chart 3
}
