% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_national_measles_coverage_different_sources.R
\name{mod_national_measles_coverage_different_sources_ui}
\alias{mod_national_measles_coverage_different_sources_ui}
\title{Create shiny module UI side for chart 6}
\usage{
mod_national_measles_coverage_different_sources_ui(id)
}
\arguments{
\item{id}{The id (string) to be used  for name-spacing in the module.}
}
\value{
A shiny UI Module
}
\description{
\code{mod_national_measles_coverage_different_sources_ui()}
Creates a shiny Module UI of National MCV coverage (\%) by different sources
currently chart 6
}
